+++ packages/orq-rc/src/orq_ai_sdk/models/createpromptop.py
@@ -150,6 +150,8 @@ class ModelParametersTypedDict(TypedDict):
     r"""The format to return the embeddings"""
     reasoning_effort: NotRequired[ReasoningEffort]
     r"""Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response."""
+    budget_tokens: NotRequired[float]
+    r"""Gives the model enhanced reasoning capabilities for complex tasks. A value of 0 disables thinking. The minimum budget tokens for thinking are 1024. The Budget Tokens should never exceed the Max Tokens parameter. Only supported by `Anthropic`"""
 
 
 class ModelParameters(BaseModel):
@@ -220,6 +222,11 @@ class ModelParameters(BaseModel):
     ] = None
     r"""Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response."""
 
+    budget_tokens: Annotated[Optional[float], pydantic.Field(alias="budgetTokens")] = (
+        None
+    )
+    r"""Gives the model enhanced reasoning capabilities for complex tasks. A value of 0 disables thinking. The minimum budget tokens for thinking are 1024. The Budget Tokens should never exceed the Max Tokens parameter. Only supported by `Anthropic`"""
+
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
         optional_fields = [
@@ -239,6 +246,7 @@ class ModelParameters(BaseModel):
             "photoRealVersion",
             "encoding_format",
             "reasoningEffort",
+            "budgetTokens",
         ]
         nullable_fields = ["responseFormat"]
         null_default_fields = []
@@ -698,6 +706,8 @@ class CreatePromptModelParametersTypedDict(TypedDict):
     r"""The format to return the embeddings"""
     reasoning_effort: NotRequired[CreatePromptReasoningEffort]
     r"""Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response."""
+    budget_tokens: NotRequired[float]
+    r"""Gives the model enhanced reasoning capabilities for complex tasks. A value of 0 disables thinking. The minimum budget tokens for thinking are 1024. The Budget Tokens should never exceed the Max Tokens parameter. Only supported by `Anthropic`"""
 
 
 class CreatePromptModelParameters(BaseModel):
@@ -771,6 +781,11 @@ class CreatePromptModelParameters(BaseModel):
     ] = None
     r"""Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response."""
 
+    budget_tokens: Annotated[Optional[float], pydantic.Field(alias="budgetTokens")] = (
+        None
+    )
+    r"""Gives the model enhanced reasoning capabilities for complex tasks. A value of 0 disables thinking. The minimum budget tokens for thinking are 1024. The Budget Tokens should never exceed the Max Tokens parameter. Only supported by `Anthropic`"""
+
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
         optional_fields = [
@@ -790,6 +805,7 @@ class CreatePromptModelParameters(BaseModel):
             "photoRealVersion",
             "encoding_format",
             "reasoningEffort",
+            "budgetTokens",
         ]
         nullable_fields = ["responseFormat"]
         null_default_fields = []